add seperate component for table pagination
